requirements-full.txt
# Full requirements for local dev and Docker/Railway
# Core framework
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-genai>=2.0.0

# Data ingestion & processing
httpx>=0.27.0
beautifulsoup4>=4.12.0
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.13.0
yfinance>=0.2.40
jugaad-data>=0.24.0
feedparser>=6.0.11

# Storage
aiosqlite>=0.20.0

# API & Frontend
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.4
python-multipart>=0.0.9
plotly>=5.22.0

# Voice
gtts>=2.5.0

# Utilities
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.1
apscheduler>=3.10.4
loguru>=0.7.2
tenacity>=8.5.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
